[refactor]: 아카이브 - 위치 이동하기 모달창 파일 분리
